microservices/generator/requirements.txt
requests==2.31.0
ruamel.yaml==0.18.5
isodate==0.6.1